Fix problem with types

+++ materials/src/refactoring/Customer.java
@@ -49,9 +49,9 @@ class Customer
 		return result;
 	}
 
-	private int amountFor(Rental each)
+	private double amountFor(Rental each)
 	{
-		int thisAmount = 0;
+		double thisAmount = 0;
 		switch (each.getMovie().getPriceCode())
 		{
 			case Movie.REGULAR:
